sim: lm32: reduce -Wno-error scope
Clean up some warnings in dv-lm32cpu, and all in sim-if, then reduce
the -Werror disable to the files that still aren't clean that now that
we require GNU make and can set variables on a per-object basis.

+++ sim/lm32/dv-lm32cpu.c
@@ -141,7 +141,7 @@ deliver_lm32cpu_interrupt (struct hw *me, void *data)
   SIM_DESC sd = hw_system (me);
   sim_cpu *cpu = STATE_CPU (sd, 0);	/* NB: fix CPU 0.  */
   address_word cia = CPU_PC_GET (cpu);
-  int interrupt = (int) data;
+  int interrupt = (uintptr_t) data;
 
 
   HW_TRACE ((me, "interrupt-check event"));
